+++ PremiumScraps/Utils/Effects.cs
@@ -385,19 +385,15 @@ namespace PremiumScraps.Utils
         }
 
         // Modified from Mrov's version
-        public static void SpawnLightningBolt(Vector3 strikePosition, bool damage = true, bool redirectInside = true, bool originInSky = true, Vector3 originOverride = default)
+        public static void SpawnLightningBolt(Vector3 strikePosition, bool damage = true, bool redirectInside = true)
         {
             LightningBoltPrefabScript localLightningBoltPrefabScript;
             var random = new System.Random(StartOfRound.Instance.randomMapSeed);
             random.Next(-32, 32); random.Next(-32, 32);
-            Vector3 origin;
-            if (originInSky)
-                origin = strikePosition + Vector3.up * 160f + new Vector3(random.Next(-32, 32), 0f, random.Next(-32, 32));
-            else
-                origin = originOverride;
-            if (redirectInside && Physics.Linecast(origin, strikePosition + Vector3.up * 0.5f, out _, StartOfRound.Instance.collidersAndRoomMaskAndDefault, QueryTriggerInteraction.Ignore))
+            var vector = strikePosition + Vector3.up * 160f + new Vector3(random.Next(-32, 32), 0f, random.Next(-32, 32));
+            if (redirectInside && Physics.Linecast(vector, strikePosition + Vector3.up * 0.5f, out _, StartOfRound.Instance.collidersAndRoomMaskAndDefault, QueryTriggerInteraction.Ignore))
             {
-                if (!Physics.Raycast(origin, strikePosition - origin, out var rayHit, 100f, StartOfRound.Instance.collidersAndRoomMaskAndDefault, QueryTriggerInteraction.Ignore))
+                if (!Physics.Raycast(vector, strikePosition - vector, out var rayHit, 100f, StartOfRound.Instance.collidersAndRoomMaskAndDefault, QueryTriggerInteraction.Ignore))
                     return;
                 strikePosition = rayHit.point;
             }
@@ -406,7 +402,7 @@ namespace PremiumScraps.Utils
             localLightningBoltPrefabScript.enabled = true;
             localLightningBoltPrefabScript.Camera = GameNetworkManager.Instance.localPlayerController.gameplayCamera;
             localLightningBoltPrefabScript.AutomaticModeSeconds = 0.2f;
-            localLightningBoltPrefabScript.Source.transform.position = origin;
+            localLightningBoltPrefabScript.Source.transform.position = vector;
             localLightningBoltPrefabScript.Destination.transform.position = strikePosition;
             localLightningBoltPrefabScript.CreateLightningBoltsNow();
             AudioSource audioSource = Object.Instantiate(stormy.targetedStrikeAudio);